Add OS command passthrough and update README

+++ SmartCLI-Submission/src/main/java/com/lpu/smartcli/commands/DeleteCommand.java
@@ -12,6 +12,11 @@ public class DeleteCommand implements Command {
             return;
         }
 
+        if (!fs.fileExists(args[0])) {
+            ErrorHandler.fileNotFound(args[0]);
+            return;
+        }
+
         fs.deleteFile(args[0]);
     }
